Checks if a binary tree is full

+++ 14-binary_tree_balance.c
@@ -45,10 +45,10 @@ int binary_tree_balance(const binary_tree_t *tree)
 		return (0);
 
 	if (tree->left != NULL)
-		left_height = calculate_tree_height(tree->left) + 1;
+		left_height = (int)calculate_tree_height(tree->left) + 1;
 
 	if (tree->right != NULL)
-		right_height = calculate_tree_height(tree->right) + 1;
+		right_height = (int)calculate_tree_height(tree->right) + 1;
 
 	return (left_height - right_height);
 }
